Renamed from PageLayout to MainLayout
